added column spec, temporary fix for #124
